# BackEnd/src/api/swagger/api/openapi.yaml
openapi: 3.0.1
info:
  title: Ecommecer docs APIs
  # description: NS Checkin system REST API
  # termsOfService: http://swagger.io/terms/
  # contact:
  #   name: Hyperledger Team
  # license:
  #   name: Apache 2.0
  #   url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://192.168.1.10:9999
  - url: http://localhost:9999
paths:
...
  /api/product/createProduct:
    post:
      tags:
        - Product
      summary: create Product
      operationId: createProduct
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                NAME:
                  type: string
                PRICE:
                  type: number
                STOCK:
                  type: number
                CD:
                  type: string
                CATEGORY_ID:
                  type: number
                PRODUCT_ATTRIBUTE:
                  type: array
                  items:
                    type: string

              example:
                image: "binary-data"
                NAME: "Product Name"
                PRICE: 20.0
                STOCK: 100
                PRODUCT_ATTRIBUTE:
                  - key: "Color"
                    value: "Blue"
                  - key: "Size"
                    value: "Medium"
      responses:
        "200":
          description: Success
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: UnAuthorized
          content: {}
        "404":
          description: Not found
          content: {}
      security:
        - JWT: []
...
components:
...
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization